added active global and all rules section
