+++ app/main/Rules/GlobalRules/style.module.css
@@ -1,0 +1,16 @@
+.topBtns{
+    border-bottom: 2px solid grey;
+    display: flex;
+    gap: 10px;
+    margin-top: -5px;
+    margin-bottom: 5px;
+}
+
+.globalBtn{
+    color: grey;
+}
+.selected{
+    border-bottom: 3px solid #22B573;
+    color: #22B573;
+    font-weight: 500;
+}
