+++ app/main/Rules/MyRules/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const MyRules = () => {
+  return (
+    <div>MyRules</div>
+  )
+}
+
+export default MyRules
