+++ app/main/Rules/SuggestedRules/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const SuggestedRules = () => {
+  return (
+    <div>SuggestedRules</div>
+  )
+}
+
+export default SuggestedRules
